+++ src/main/java/net/minecraftforge/common/ForgeMod.java
@@ -28,7 +28,6 @@ import net.minecraft.util.SoundEvent;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.world.storage.SaveHandler;
 import net.minecraft.world.storage.WorldInfo;
-import net.minecraftforge.client.gui.DefaultConfigScreen;
 import net.minecraftforge.common.crafting.*;
 import net.minecraftforge.common.crafting.conditions.*;
 import net.minecraftforge.common.data.ForgeBlockTagsProvider;
@@ -42,6 +41,7 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fml.*;
+import net.minecraftforge.fml.client.gui.screen.DefaultConfigScreen;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
